Cherry pick a test fix.
